third_party/blink/web_tests/http/tests/xmlhttprequest/get-dangerous-headers-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
CONSOLE ERROR: Refused to get unsafe header "Set-Cookie"
CONSOLE ERROR: Refused to get unsafe header "set-cookie"
CONSOLE ERROR: Refused to get unsafe header "Set-Cookie2"
CONSOLE ERROR: Refused to get unsafe header "set-cookie2"
Test that getResponseHeader and getAllResponseHeaders cannot be used to get the cookie header fields.

PASS
